# lcm-engine/lcm_engine/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: |
    The Life-Cycle Management Engine API specification for the Gaia-X project.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: LCM Engine API
  version: 2.3.0
servers:
- description: Testing LCM Engine server 1
  url: https://demo.gxfs.xlab.si
- description: Testing LCM Engine server 2
  url: https://orchestrator.gxfs.xlab.si
- description: Local development
  url: http://localhost:8080
...
paths:
...
  /workspace/{workspaceId}/project:
    get:
      operationId: list_workspace_projects
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Project'
                type: array
          description: A list of projects
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: List projects in a workspace
      tags:
      - project
      x-openapi-router-controller: lcm_engine.controllers.project_controller
    post:
      description: "Secrets applied to the workspace the project is in are only applied\
        \ on creation.\nTo modify secrets, create a new project.\n"
      operationId: create_workspace_project
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/ProjectBody'
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityReference'
          description: A reference to the created project
        "400":
          content:
            application/json:
              example:
                msg: Request has incorrect schema or values
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Create a new project in the workspace (async)
      tags:
      - project
      x-openapi-router-controller: lcm_engine.controllers.project_controller
  /workspace/{workspaceId}/project/{projectId}:
    delete:
      operationId: delete_workspace_project
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - explode: false
        in: path
        name: projectId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Project deleted
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Delete the project
      tags:
      - project
      x-openapi-router-controller: lcm_engine.controllers.project_controller
    get:
      operationId: describe_workspace_project
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - explode: false
        in: path
        name: projectId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: A project
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Describe the project
      tags:
      - project
      x-openapi-router-controller: lcm_engine.controllers.project_controller
    patch:
      operationId: update_workspace_project
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - explode: false
        in: path
        name: projectId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/ProjectBody'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: The updated project
        "400":
          content:
            application/json:
              example:
                msg: Request has incorrect schema or values
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Update the project
      tags:
      - project
      x-openapi-router-controller: lcm_engine.controllers.project_controller
    put:
      operationId: replace_workspace_project
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - explode: false
        in: path
        name: projectId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/ProjectBody'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
          description: The updated project
        "400":
          content:
            application/json:
              example:
                msg: Request has incorrect schema or values
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad request
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Replace the project
      tags:
      - project
      x-openapi-router-controller: lcm_engine.controllers.project_controller
    x-extension-marker: true
  /workspace/{workspaceId}/project/{projectId}/creationStatus:
    get:
      operationId: describe_workspace_project_status
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - explode: false
        in: path
        name: projectId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityCreationStatus'
          description: The project creation status
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Get the project creation status
      tags:
      - project
      x-openapi-router-controller: lcm_engine.controllers.project_controller
  /workspace/{workspaceId}/project/{projectId}/debugPackage:
    get:
      operationId: create_workspace_project_debug_package
      parameters:
      - explode: false
        in: path
        name: workspaceId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - explode: false
        in: path
        name: projectId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: An authorization header
        example: john.doe@example.com
        explode: false
        in: header
        name: X-Forwarded-User
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/zip:
              schema:
                format: binary
                type: string
          description: A debug archive
        "401":
          content:
            application/json:
              example:
                msg: User is not authorized to perform this action
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        "404":
          content:
            application/json:
              example:
                msg: Given resource was not found
              schema:
                $ref: '#/components/schemas/Error'
          description: Not found
      summary: Get a package for local debugging
      tags:
      - project
      x-openapi-router-controller: lcm_engine.controllers.project_controller
...
components:
...
  requestBodies:
...
    ProjectBody:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Project'
      description: Project specification object
      required: true
...
  schemas:
...
    EntityReference:
      additionalProperties: false
      description: A reference to another entity
      example:
        id: 0
      properties:
        id:
          format: int64
          title: id
          type: integer
      required:
      - id
      title: EntityReference
      type: object
    EntityCreationStatus:
      additionalProperties: false
      description: The status of an entity creation request
      example:
        finished: true
        status: status
      properties:
        finished:
          title: finished
          type: boolean
        status:
          title: status
          type: string
      required:
      - finished
      - status
      title: EntityCreationStatus
      type: object
...
    Project:
      additionalProperties: false
      description: A project
      example:
        workspace: 8
        kind: si.xlab.tosca
        name: my tosca lcm service
        id: 2
        csar: csar
      properties:
        id:
          example: 2
          format: int64
          readOnly: true
          title: id
          type: integer
        name:
          example: my tosca lcm service
          title: name
          type: string
        workspace:
          example: 8
          format: int64
          readOnly: true
          title: workspace
          type: integer
        csar:
          description: Base64-encoded ZIP file of the CSAR. The service template must
            be top-level
          format: byte
          title: csar
          type: string
          writeOnly: true
        kind:
          description: "Kind of a project, such as TOSCA or Terraform"
          example: si.xlab.tosca
          title: kind
          type: string
      required:
      - csar
      - id
      - kind
      - name
      - workspace
      title: Project
      type: object
...
    Error:
      additionalProperties: false
      description: An error
      properties:
        msg:
          title: msg
          type: string
      required:
      - msg
      title: Error
      type: object
...